+++ common/src/lib.rs
@@ -1,1 +1,3 @@
 pub mod user;
+pub mod command;
+pub mod error;
