+++ podman-pilot/src/podman.rs
@@ -5,6 +5,9 @@ use flakes::user::User;
 use nix::unistd;
 use spinoff::{spinners, Color, Spinner};
 use std::fs;
+use flakes::error::FlakeError;
+use tempfile::tempfile;
+use std::io::{Write, Read};
 use std::fs::File;
 use std::io::Seek;
 use std::io::SeekFrom;
